fix ci and refactor common util
