+++ e2e/cli.go
@@ -28,12 +28,6 @@ import (
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
 	"github.com/onsi/gomega/gexec"
-	k8sruntime "k8s.io/apimachinery/pkg/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/config"
-
-	oamcore "github.com/oam-dev/kubevela/apis/core.oam.dev"
 )
 
 var rudrPath = GetCliBinary()
@@ -113,26 +107,6 @@ func InteractiveExec(cli string, consoleFn func(*expect.Console)) (string, error
 		return string(output), err
 	}
 	return string(s.Out.Contents()) + string(s.Err.Contents()), nil
-}
-
-func NewK8sClient() (client.Client, error) {
-	conf, err := config.GetConfig()
-	if err != nil {
-		return nil, err
-	}
-	scheme := k8sruntime.NewScheme()
-	if err := clientgoscheme.AddToScheme(scheme); err != nil {
-		return nil, err
-	}
-	if err := oamcore.AddToScheme(scheme); err != nil {
-		return nil, err
-	}
-
-	k8sclient, err := client.New(conf, client.Options{Scheme: scheme})
-	if err != nil {
-		return nil, err
-	}
-	return k8sclient, nil
 }
 
 func BeforeSuit() {
