+++ pkg/utils/common/common.go
@@ -48,6 +48,7 @@ import (
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	ocmclusterv1alpha1 "open-cluster-management.io/api/cluster/v1alpha1"
 	ocmworkv1 "open-cluster-management.io/api/work/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 
 	oamcore "github.com/oam-dev/kubevela/apis/core.oam.dev"
@@ -281,4 +282,25 @@ func SetSpecObjIntoUnstructuredObj(spec interface{}, u *unstructured.Unstructure
 	}
 	_ = unstructured.SetNestedMap(u.Object, data, "spec")
 	return nil
+}
+
+// NewK8sClient init a local k8s client which add oamcore scheme
+func NewK8sClient() (client.Client, error) {
+	conf, err := config.GetConfig()
+	if err != nil {
+		return nil, err
+	}
+	scheme := k8sruntime.NewScheme()
+	if err := clientgoscheme.AddToScheme(scheme); err != nil {
+		return nil, err
+	}
+	if err := oamcore.AddToScheme(scheme); err != nil {
+		return nil, err
+	}
+
+	k8sClient, err := client.New(conf, client.Options{Scheme: scheme})
+	if err != nil {
+		return nil, err
+	}
+	return k8sClient, nil
 }
